benchmark/pyproject.toml
[project]
name = "murphysig-benchmark"
version = "0.1.0"
description = "Empirical benchmark: do MurphySig signatures measurably change how AI models treat code?"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that make real API calls",
]
